Support sequences without using maximum(:id)
Minifacture is unable to run sequences for models that don't use an
integer for their primary keys (e.g. UUIDs). To solve this, we can have
each factory maintain a count that is incremented as `build` is called.
This count is also inherited from parents to simulate how sequences
would work in cases of single table inheritance.

Signed-off-by: David Celis <me@davidcel.is>